+++ src/main/java/bjs2_57854/Image.java
@@ -1,4 +1,4 @@
-package school.faang;
+package bjs2_57854;
 
 public record Image(String name, String description) {
 }
