+++ src/test/java/bjs2_57854/FilterProcessorTest.java
@@ -1,12 +1,10 @@
-package school.faang;
+package bjs2_57854;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import java.util.function.Function;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 class FilterProcessorTest {
     Image testImage;
